+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to the "Hydro Helper" extension will be documented in this f
 
 ## v0.1.2
 
+- **Add webviewColumn setting to configure webview display position**
 - Refactor score and status checks to use strict undefined comparisons
 
 Want to see the previous log? [Click here](https://github.com/langningchen/hydro-helper/commits/main/CHANGELOG.md)